Fix `CFileWriter::Truncate()`: call `fflush()` immediately before `ftruncate()`.
The buffer must be empty on `ftruncate()` because we are operating on the
underlying file descriptor. Calling `fseek()` between `fflush()` and
`ftruncate()` could have caused the C library to reload the buffer, and the
buffered data could be returned by the `Reader` returned by `ReadMode()` even
after the underlying file descriptor has been truncated.

The bad scenario is:
* `Write(...)`.
* `Flush()`.
* `Seek(0)`.
* `Truncate(0)`.
* `ReadMode()`.
* `Reader::Pull()`.

PiperOrigin-RevId: 647247400