Fix conversion of `InvokerType` to `InvokerType::ConstResult`.
A template with a defaulted `DependentConstResult` parameter, which was thought
of solving the error at `operator void()` when `ConstResult` is `void`, does not
work as intended, because the caller can attempt conversion to a different type,
causing `DependentConstResult` to be something else than `ConstResult`, in which
case the body might not compile.

C++20 `requires` cannot be assumed to be available yet either.

So the solution for making conversion operators conditionally present is
deriving from a helper template with separate specializations depending on
whether the function is const-callable.

Also, there is no need for `kIsConstCallable`, because conditional presence of
`ConstResult` is SFINAE-friendly.

PiperOrigin-RevId: 648981317