Let `MemoryEstimator` support `std::unique_ptr<T, Deleter>` with a non-default `Deleter`, as well as `std::shared_ptr<T[size]>` (`std::unique_ptr` does not support `T[size]`).
Remove an explicit override for `std::shared_ptr<void>`. Still explicitly
exclude it from the main override, as well as `std::unique_ptr<void>`,
`std::unique_ptr<T[]>`, and `std::shared_ptr<T[]>`.

It is infeasible to have a good estimation for them (for `void` the type is
completely unknown, for `T[]` the number of elements is unknown).

Excluding `void` is necessary for the main override to compile. Not providing
separate overrides lets encountering them register them as unknown types,
which might encourage their usage to change their type to something more
introspectable if a better memory estimation is desired.

PiperOrigin-RevId: 652426353