Move the responsibility for calling `Chain::EnsureHasHere()` (deallocating the block pointer array before writing inline data) from functions calling `Chain::Initialize()` to `Chain::Initialize()` functions themselves.
This frees the block pointer array only when writing inline data, not when block
pointers are needed again, reusing the allocation of the block pointer array, at
the cost of a redundant comparison in constructors where blocks are definitely
not allocated, and retaining memory for the block pointer array.

This fixes `Chain::operator=(const Chain&)` when allocated block pointers are
being replaced with inline data: `Chain::EnsureHasHere()` was missing.

Add assertions about preconditions of `Chain::Initialize()`.

PiperOrigin-RevId: 653193048