Add `InitializerTargetRefT<T>`.
It is like `InitializerTargetT<T>`, but if the object is already constructed,
then it is the corresponding reference type instead of the value type.

This allows to avoid copying the object if a const reference to it can be
sufficient under some conditions. `Initializer<const T&>::Reference()` returns
`const T&`.

Allow providing const variants of `operator()` and
`RiegeliTo{ChainBlock,Cord,ExternalData,ExternalStorage}()`. If only a mutable
variant is provided and the object is passed by const or lvalue reference, the
function will be called on a mutable copy of the object.

Remove the optimization of constructing the external object already wrapped in
`ObjectForCordSubstr` which avoided moving it. This simplifies source code,
especially given planned changes here. An object with stable substrings is
likely to be efficiently movable, and it will be moved again by
`absl::MakeCordByExternal()` anyway.

PiperOrigin-RevId: 655551560