Add `TypeErasedRef` to replace manual taking of the address and casting to `void*` and casting back.
This is tricky enough to get right: handling references required
`std::remove_reference_t`, handling const references required `const_cast`,
and this did not compile for references to functions which do not support
`const_cast`.

References to functions could be handled by decaying them to pointers to
functions, but it is more efficient to store the function pointer inline,
avoiding an indirection.

Sadly, in general rvalue references which would fit in pointer space cannot
be inlined, because some callers require preserving the address (so that it
refers to an object which is still alive instead of a local temporary).

PiperOrigin-RevId: 658785105