Use `readlink()` on `/proc/self/fd/$FD` for recovering `filename()` from a fd.
The cost is +3.5 µs for creating e.g. `StdOut` (it took 2 µs before).
I feel that better failure messages which embed `filename()` are worth it.

On Windows the placeholder is always used. I believe `CONIN$` etc. are
misleading if standard streams are redirected.

Remove `set_assumed_filename()` options. The automatically recovered filename is
good enough, at least on Linux, so there should be little reason to override it.

PiperOrigin-RevId: 658803752