Add `ByteFill`, representing a number of bytes filled with a single value, replacing `{{,Backward}Writer,DigesterHandle}::Write{Zeros,Chars,Bytes}()` and `{Array,Chain,Cord}OfZeros()`, and providing iteration over blocks which are convertible to `ExternalRef`.
This makes it easier to combine writing repetitions with writing other objects
in a single variadic call to `Write()` or `DigestFrom()`.

This also makes it easier to implement materializing the sequence by iterating
over blocks.

`WriteSlow(ByteFill)` is virtual, replacing virtual `WriteZerosSlow()` and
non-virtual `WriteCharsSlow()`. This means that filling with a value other than
zero is customizable too, at the cost of having to customize for any value even
if only filling with zeros is customizable in an interesting way.

PiperOrigin-RevId: 658855384