Fix `Initializer<T>::UniquePtr()` for overaligned types by using the same lookup rules of `operator new` as for the `new` expression.
If aligned `operator new` is supported, for allocation of overaligned types look
up in order:
* `T::operator new(size_t, std::align_val_t)`
* `T::operator new(size_t)` (in this case C++ assumes that it understands that
  the type requires a particular alignment)
* `::operator new(size_t, std::align_val_t)`

If aligned `operator new` is not supported and `T::operator new(size_t)` is not
defined, require that the alignment is at most `alignof(max_align_t)`.

If the type is not overaligned and only
`T::operator new(size_t, std::align_val_t)` is supported, fail (it hides
`T::operator new(size_t)`).

PiperOrigin-RevId: 670488188