Remove `ReferenceOrCheapValue`; always pass by reference instead.
Rationale: in rare cases the identity of the object passed by rvalue reference
matters, as the constructor or function can return a value retaining the address
of that object. OTOH savings from passing by value are minor, because in most
cases the affected code is inlined and the object containing a
`ReferenceOrCheapValue` is never materialized. It is safer to avoid a potential
trap when function arguments are not forwarded perfectly.

Also, reduce the number of explicit specializations of `MakerType` (for 0 and 1
arguments instead of up to 4) ad `Invoker{,Const}Base` (for 0 arguments instead
of up to 4). These specializations are less important when these types rarely
contain non-references and `ABSL_ATTRIBUTE_NO_UNIQUE_ADDRESS` has fewer chances
to apply.

This indirectly fixes existing bugs:

* `ReferenceOrCheapValue` was wrong for some non-copyable types because
  for some reason `std::is_trivially_copyable` can be `true` even when
  `std::is_copy_constructible` is `false`.

* `InvokerConstBase` had typos in cases for 3 and 4 parameters.

Also, fix `InitializerTarget` for `Initializer<T>::AllowExplicit`.

PiperOrigin-RevId: 672614013