Make `Any`, `Dependency`, and `StableDependency` unconditionally nullability compatible.
Abseil is changing the way it marks classes being nullability compatible from
the `absl_nullability_compatible` tag type to the `ABSL_NULLABILITY_COMPATIBLE`
attribute (which is an alias for Clang's `_Nullable` attribute). See
[here](https://github.com/abseil/abseil-cpp/blob/ba5fd0979b4e74bd4d1b8da1d84347173bd9f17f/absl/base/nullability.h#L129)
for details.

This will allow `_Nullable` and `_Nonnull` attributes to be added to the
`absl::Nullable` and `absl::Nonnull` annotations, and this in turn allows Clang
to diagnose some nullability errors at compile time through the `-Wnonnull` and
`-Wnullability` warnings.

However, Clang does not consider the `_Nullable` class attribute to be inherited
by derived classes, so the `ConditionallyAbslNullabilityCompatible` mechanism no
longer works. Instead, we mark the relevant derived classes as unconditionally
nullability compatible; in practice, most specializations of these classes are
in fact nullability compatible.

PiperOrigin-RevId: 678670335