Let `InitializerTargetT<InvokerType>` decay the result to a value type. `InitializerTargetRefT<InvokerType>` keeps the original type which can be a reference.
This allows changing a function wrapped by `riegeli::Invoker()` to return a
reference instead of a value without breaking the target type deduced by
`InitializerTargetT<InvokerType>`.

Deducing the target type from `MakerTypeFor` or `Initializer` still does not
decay the target type. These types are considered explicit markers that the
exact specified type is intended, even if it is a reference.

Make `InitializerTarget` and `InitializerTargetRef` more precise for unusable
types in a SFINAE-friendly way: they exist but without a `type` member if their
parameter cannot be used.

To support that, rename `InvokerResult` to `InvokerTargetRef`, and add
`InvokerTarget` and `MakerTarget`.

PiperOrigin-RevId: 690661548