Add `Initializer` constructors from a parameter describing an object involving a different but compatible target for `MakerTypeFor` and `Initializer`.
Previously only a direct parameter convertible to `T` and `InvokerType` handled
different but compatible targets. For `MakerTypeFor` and `Initializer` that
worked inconsistently depending on whether conversion was a user defined
conversion or a built-in conversion, because C++ looks for only one user defined
conversion at a time, surrounded by built-in conversions.

PiperOrigin-RevId: 690742784