Let `{Linear,Chunked}SortedStringSet::contains()` optionally provide the index of `element` in the set, or to the index where it would be inserted.
This is useful if there is a parallel sequence corresponding to elements of the
set, e.g. to implement a map or a BTree node.

This required rearranging the internal representation, so that
`ChunkedSortedStringSet` maintains the cumulative sizes of all chunks,
not only of the last one. This turned out to lead to code simplifications
by using `absl::InlinedVector<Chunk, 1>` instead of storing the first chunk
separately.

Also, optimize `ChunkedSortedStringSet::contains()` by calling the new
`LinearSortedStringSet::contains_skip_first()` when it is known that the first
element does not match.

Made `{Linear,Chunked}SortedStringSet::Builder::Build()` revert to an empty
state instead of leaving it indeterminate, and remove rvalue qualification.
This behavior of `LinearSortedStringSet::Builder::Build()` is useful in
`ChunkedSortedStringSet::Builder::Build()`.

PiperOrigin-RevId: 691483158