Make `riegeli::Compare()` working for `absl::string_view`, and use it instead of `absl::string_view::compare()`, to use more meaningful types.
In theory an early conversion from `int` to `StrongOrdering` might cause
unnecessary conditionals, if the result is then only compared against 0.

But `BinarySearch()` performs this conversion anyway (and if that is inlined
together with the test function, then it might be optimized away), and in other
cases there should be little difference.

PiperOrigin-RevId: 691483653