Do not bother with considering the case when a flat array in memory exceeds `std::streamsize` range.
`std::streamsize` is a signed counterpart of `size_t`. In all modern systems
`size_t` has the same size as a pointer, and there is no way to create an array
which occupies over half of the address space.

Pointer subtraction would be undefined for such an array anyway (and unfixable
portably), so `available()` already would have undefined behavior.

PiperOrigin-RevId: 692479150