Support digest conversions also from arrays of integers of smaller sizes to integers or arrays of integers of larger sizes.
To prevent infinite recursion, in these cases further conversions using
`DigestConverterImpl` are not considered, only types which can be natively
explicitly converted.

This is useful to express the result of `HighwayHash128Digester` as
`absl::uint128` instead of `std::array<uint64_t, 2>`, or to express the result
of OpenSSL-based digesters as `std::array<uint64_t, size>` instead of
`std::array<char, size * sizeof(uint64_t)>`.

Also, ensure that if a digest is already natively convertible to the target
type, then `DigestConverterImpl` does not consider any other conversion first,
which would be ambiguous given that the native conversion is available too.

PiperOrigin-RevId: 693311655