Unify naming of parameters in `AbslStringify()`, `operator<<`, `AbslFormatFlush()`, and `Sink::Append(length, fill)`.
Use `src` and `dest`, especially as the parameter order differs from the regular
Riegeli convention of putting the output parameter last.

Avoid `self` in extension points using FTADLE, because in some contexts they are
defined as global functions. `self` is reserved for cases where something like
`this` is meant but cannot be written as such because we are in a helper object,
like a visitor for `absl::variant`.

PiperOrigin-RevId: 693377499