Make `riegeli::invoke()`, `is_invocable`, and `invoke_result{,_t}` public, replacing internal `IsCallable` and explicit `#if` / `decltype()`. They correspond to C++17 features with the same names. They are getting enough usages.
Remove `InvokerType::{,Const}Result` type aliases. Use
`invoke_result_t<Function&&, Args&&...>` and
`invoke_result_t<const Function&, const Args&...>` instead for getting the type,
or `is_invocable` for checking applicability of `InvokerType`. This avoids
instantiating the type to check its applicability, which is prone to cycles in
template instantiations. They were not observed for `InvokerType` yet, but this
change is being done just in case, and for consistency with `MakerTypeFor`.

Split `InvokerType` internals such that only conversion operators are included
conditionally. This is required for conversion operators (SFINAE is applicable
to conversion operators in some cases, but not when the target type exists
conditionally). For other functions (they will come) and for the constructor,
SFINAE constraints yield better error messages: that the function is not
applicable because of a particular constraint, rather than that it is unknown.

Add SFINAE constraints that the index is valid for `InvokerType::arg<index>()`.

PiperOrigin-RevId: 694004588