Rename `InitializerTarget{,Ref}{,T}` to `Target{,Ref}{,T}`.
They are related to `Initializer` but they are applicable in their own ways also
to everything `Initializer` can be constructed from: already constructed
objects, `riegeli::Maker()`, and `riegeli::Invoker()`.

In the context of Riegeli this is an important core concept.

The names are quite long. Related names will come, which would get even longer.

PiperOrigin-RevId: 694485013