Fix `AnyRef` and `AnyInitializer` constructors to exclude not only exact conflicts with the move constructor, but also conflicts with materializing the given object from `Initializer` etc.
This way constructing `AnyRef<Handle>` from `Initializer<AnyRef<Handle>>` is
done by materializing the `AnyRef<Handle>`, instead of wrapping a reference to
it. The latter might not work anyway because it could store a reference to an
object which is getting destroyed too soon, which is caught by
`ABSL_ATTRIBUTE_LIFETIME_BOUND`.

Cosmetics: remove old-style `absl_nullability_compatible` tag which does not
have any effect since
https://github.com/abseil/abseil-cpp/commit/485f2be7c1d2f1a2ad2d89f6aef308dd7bfede8d
and was left here by mistake.
PiperOrigin-RevId: 697950999