Prefer `absl::string_view::substr()` to `absl::string_view` constructor from pointer and length, adding `RIEGELI_ASSUME()` for optimization.
It used to be the case that `absl::string_view` constructor from pointer and
length had no bounds checking, while `absl::string_view::substr()` checked that
the starting position is in range (throwing an exception otherwise) and that the
length is not larger than the remaining size (reducing the length otherwise).

Nowadays `absl::string_view` constructor from pointer and length
checks that the length fits in `ptrdiff_t`, also in opt mode. Hence we use
`absl::string_view::substr()` which does not need this check, and if we are sure
that the position and length fit, a `RIEGELI_ASSUME()` allows the compiler
to skip runtime checks.

Skip following an impossible condition in
`LinearSortedStringSet::ContainsImpl()`:
it was checked that `common_length < found_prefix().size()`,
hence `!found_middle.empty()`.

Add missing `ABSL_ATTRIBUTE_LIFETIME_BOUND` annotations.

PiperOrigin-RevId: 707059598