Clean up how a templated constructor or assignment exclude parameter types which would conflict with the copy or move constructor or assignment.
`NotSelfCopy<Self, Args...>` captures the proper constraint, documenting the
purpose and hiding a subtle implementation.

The constraint excludes not only `Self` but also classes derived from `Self`,
so that the constraint can be put in a base class without mentioning derived
classes, yet it will be triggered when a derived class inherits the constructor
or assignment.

PiperOrigin-RevId: 707094234