Let constructing `AnyRef<Handle>` from `Any<Handle, inline_size, inline_align>` adopt the state if possible instead of forming a reference.
This is done by extending `any_internal::IsAny` to accept also rvalue references
to `Any` and `AnyRef`, so that private `Initialize()` functions recognize also
such rvalue references as types to consider adopting the state from.

This required adapting these functions, as well as
`AnyInitializer::ConstructMethod()` which also uses `any_internal::IsAny`,
to apply `std::remove_reference_t` for accessing static members of the `Any`
or `AnyRef`, because rvalue references are not transparent in this case.

PiperOrigin-RevId: 713281200