Parameterize `ExternalRef::StorageWhole` by `TargetRefT<Arg>` instead of by `Arg&&`.
The implementation of `StorageWhole` depends only on the target type, not on
its initializer, because the object needs to be constructed unconditionaly for
conversion to `absl::string_view`; the `StorageWhole` keeps the constructed
object rather than its initializer.

While this is inconsistent with `ExternalRef::StorageSubstr`, this allows
to skip the details of `MakerTypeFor` and pass only the target type if the
`StorageWhole` is passed explicitly — which is very rare, hence regular code
is unaffected.

PiperOrigin-RevId: 715287089