Add `StringRef`, `BytesRef`, and `CStringRef` to support `std::string_view` when it is a different type than `absl::string_view`. (Sadly, making `absl::string_view` convertible from `std::string_view` would lead to ambiguities e.g. regarding `operator==`.)
`StringRef` stores an `absl::string_view`, possibly converted from
`std::string_view`.

`BytesRef` stores an `absl::string_view`, possibly converted from
`std::string_view` or `absl::Span<const char>`. It replaces
`SupportsToStringView` with `ToStringView()`, but it is a type,
rather than a type predicate associated with a function, and thus
does not require making templates.

`CStringRef` stores a pointer to a C-style NUL-terminated string, or `nullptr`.
The string is copied unless it is known to end with NUL.

They are intended for function parameters when the implementation needs
`absl::string_view` or a C-style NUL-terminated string, and the caller might
have other representations of the string.

Use `CStringRef` in `Owned{Fd,CFile}::Open{,At}()` to reduce the number of
overloads. Also for Windows: while it does not need NUL-termination because
the strings are converted to `std::wstring`, the resulting performance penalty
is dwarfed by the cost of opening the file, and a unified API is nicer.

Make conversions to `absl::string_view` implicit, as is customary.

Make `CompactString` constructible/assignable from and comparable with
`BytesRef` instead of only `absl::string_view`.
Add `CompactString::{at,front,back}()` to complement `operator[]`.

PiperOrigin-RevId: 717516642