Optimize `Any::GetIf()` to check the type by a simple inlined comparison. Accessing the stored object in case of a match still involves an indirect call. This way a sequence of `GetIf()` calls to find the type has only one indirect call.
All that applies to `AnyRef` too.

To enable this, converting one `Any` to another is done differently if the
storage cannot be adopted as is. Instead of allocating or referring to the
whole `Any`, the storage is moved to a heap allocation (if the target must be
independent), or an non-owning pointer is formed (if the target can refer to the
source).

This also has an advantage that indirection through the source `Any` is
performed eagerly instead of on every virtual access.

This has a less desirable side effect of emitting more code if objects are
stored inline, because they need to support moving to a heap allocation and
forming a non-owning pointer. The code is accessed only if the new functionality
is used.

Converting from another `Any` no longer supports taking an arbitrary
`std::unique_ptr<Any, Deleter>`, but only `ClosingPtrType<Any>`. The source
`Any` with the `std::unique_ptr` no longer exists, so a non-trivial deleter
could not be called later.

Add `Any::type_id()` to enable custom type dispatching mechanisms.
Add `Any::Holds<Manager>()` as a more efficient variant of
`Any::GetIf<Manager>() != nullptr`. Remove `Any::GetIf(TypeId)` which served
only to enable `Any` conversions and was not used directly.

Remove `Dependency::GetIf<Manager>()` and `Dependency::GetIf(TypeId)`. They
served only to enable `Any` conversions and were not used directly.

The new design requires `GetIf()` to answer to a single type, as opposed to
potentially accepting various levels of wrapping as alternatives, like making
`std::unique_ptr<T>` recognizable also as `T`. The latter was once considered
but never actually implemented.

Tweak the definition of comparison of `Any<Handle>` against `nullptr`: if
`Handle` is comparable against `nullptr`, just delegate to that, without
checking for an empty `Any<Handle>` beforehand. An empty `Any<Handle>` has a
sentinel `Handle`, and if it is comparable against `nullptr`, then normally it
compares as `true`, which is definitely the case e.g. for raw pointers.

Call `Methods::move()` on the source rather than on the destination, and make it
responsible for setting the methods pointer of the destination, for consistency
with new methods.

Other cosmetic changes.

PiperOrigin-RevId: 736451619