Do not call `Py_DECREF()` from `FreeStaticObjects()`.
It is unsafe to call Python API functions from a function registered with `Py_AtExit()`, and this actually breaks in Python 3.12.

`FreeStaticObjects()` still sets pointers to static objects to `nullptr`, so that in the case of reinitializing the Python interpreter they are created again.

It is unclear for now whether the objects actually leak on interpreter shutdown, and if so, how to delete them.

PiperOrigin-RevId: 747810493