Fix UB in `TypeErasedRef`: represent both data pointers and function pointers by casting them to `void*` instead of using a union.
The old way caused UB in a null pointer check, because the check used a third
union member to examine the representation of either kind of pointers. Examining
the representation by reading an inactive union member is actually invalid.

While it would be possible to use only a byte array in the representation,
copying a data pointer or a function pointer (potentially differently sized)
to and from the byte array, it is simpler and it should be acceptable to rely on
the assumption that function pointers can be `reinterpret_cast` to and from data
pointers. This is not fully guaranteed but very common nowadays. In particular
this is true on all POSIX systems (`dlsym()` API relies on that) and on Windows.
DOS memory models which do not support this are a thing of the past.

Also, add missing `nullptr` comparisons with swapped arguments and with `!=`
before C++20.

PiperOrigin-RevId: 754048649