Move internal type traits and functions dealing with invoking actions to a separate header. This avoids `serialized_message_rewriter` reaching to an internal namespace of `serialized_message_reader`.
Reject actions invocable both with and without context as ambiguous. This might
arise if the action accepts an unconstrained list of arguments, and guessing
wrong whether the context is expected can lead to errors further down.

Move field copying from `serialized_message_writer.cc` to a new
`serialized_message_rewriter.cc`. It is used only in `SerializedMessageRewriter`
and it is unlikely to ever be useful outside it.

In `SerializedMessage{Reader,Rewriter}`, do not mention each time that the
action is invoked with `Context&` as the last argument. This is stated once,
and the parameter is optional (always missing if `Context` is `void`).

In `SerializedMessage{Reader,Rewriter}`, do not mention again that an action
returning a non-OK status causes an early exit. Especially as this is quite
obvious.

Reorder libraries to `writer`, `backward_writer`, `reader`, and `rewriter`,
so that two similar classes (`reader` and `rewriter`) are next to each other.

Mark some functions private.

PiperOrigin-RevId: 758750558