Optimize copying unchanged fields by `SerializedMessageRewriter`: copy the section of the source in bulk instead of rewriting fields one by one, as long as the source `Reader::SupportsRandomAccess()`.
This allows `Chain` or `absl::Cord` memory sharing for a whole sequence of
unchanged fields, and avoids recoding the field tag and possibly the value.

Make `LimitingReaderBase& src` parameters in
`SerializedMessage{Reader,Rewriter}` actions and `SerializedMessageWriter& dest`
parameters in `SerializedMessageRewriter` actions optional. This is useful
to avoid duplicating field decoding logic in that optimization (because
`SerializedMessageRewriter` actions need to access `src` for positions even for
already parsed fields), and also to make action lambdas shorter by removing
parameters which would be ignored.

PiperOrigin-RevId: 761059229