Ensure that reopening a file `Reader`/`Writer` with the same filename works even if the filename is taken from the `filename()` call on the same object, which is an `absl::string_view` which gets invalidated when the file handle gets reset.
This is done by converting the filename to `CompactString` early, and then
moving it to its target place, instead of passing a `PathRef` around.

Some file handle member functions, most notably `Open()`, get overloads with
`CompactString` filenames. Otherwise the public API is unchanged; the overloads
are normally used only by file `Reader`/`Writer` classes.

Add `CompactString::ForCStr()` as a single place which allocates a
`CompactString` with a room for NUL, for a more efficient `c_str()`.

PiperOrigin-RevId: 761882959