Rely on C++17: assume `__cpp_guaranteed_copy_elision`.
This simplified several places where workarounds were needed for constructing
immovable objects in the right place.

Remove `{Initializer,MakerType,MakerTypeFor,InvokerType}::ConstructAt()`. It
existed only for constructing immovable objects in-place. Since C++17 this can
be achieved with a plain placement `new` with an expression returning the given
object by value. Or more commonly, by combining construction with allocation in
a `new` expression.

In `riegeli::Global()`, remove the ability to specify an `Initializer` instead
of an already constructed object. It existed for constructing immovable objects
or for avoiding a move. Since C++17 this is achieved automatically when
returning by value.

PiperOrigin-RevId: 765658050