BEGIN_PUBLIC
Fix `Lz4Reader` with `Options::concatenate()`: if the frame is done, the next
time the header will be read, no matter whether the current `Read()` call
spanned frame boundaries (this was working) or finished at the frame boundary
(this was not).

Cosmetics:
* Reorder member variables in `{Framed,Hadoop}SnappyReader` for consistency
  with other decompressors.
* In `XzReader`, extract the logic of detecting truncation at close to a
  separate function.

EBD_PUBLIC

In `reader_test`, exercise `{Framed,Hadoop}SnappyReader` with concatenated
streams. They effectively always work in concatenate mode because their format
does not have an end of stream marker.

PiperOrigin-RevId: 772555411