Let `ExternalRef` no longer avoid embedding wasteful objects.
The canonical example of a wasteful object is a `std::string` where the capacity
is much larger than the size. Avoiding embedding them was done by switching to
copying the data.

The rationale for avoiding was that embedding wasteful objects risks producing
objects consuming `ExternalRef` (like `Cord` or `Chain`) which pin much more
memory than the size of their logical data. This is hard to detect and hard to
fix afterwards (there is no `shrink_to_fit()` on the result).

Unfortunately avoiding embedding wasteful objects has downsides too:

* When converting a `Chain` to `Cord`, if the `Chain` contains blocks which
  refer to consecutive substrings of a single large block, the blocks were
  considered wasteful and unnecessarily copied rather than embedded.

  Techically they are indeed wasteful when considered in isolation, but all
  blocks taken together pin as much memory as they need.

  An example of such a `Chain` is the result of `RecordWriter` with compression
  turned off, containing large flat records. This includes `RecordWriter` with
  `CordWriter`, where `Chain` is an internal data structure in
  `SimpleChunkEncoder`.

* If the resulting `Chain` or `Cord` is short-lived, then embedding a wasteful
  object is fine.

* If the original object can be shared (rather than moved) and outlives the
  resulting `Chain` or `Cord`, then embedding a wasteful object is fine.

* Even if a caller noticed the issue and determined that embedding a wasteful
  object is fine, there is no good way to override that. Even if `ExternalRef`
  supported some alternative API which forced embedding wasteful objects, usage
  of `ExternalRef` often happens inside other abstractions, and propagating the
  option would be problematic.

The downside of the change is a risk of wasting memory for longer time.

If a caller notices the opposite issue and determines that embedding a wasteful
object should be avoided, and it controls a place where the object is passed,
it can explicitly check whether the object is wasteful (e.g. by
`riegeli::Wasteful(riegeli::EstimateMemorySimplified(object), size)`)
and pass `absl::string_view` instead of the object in this case.

PiperOrigin-RevId: 775530406