Shorten names involving `AbslStringify` by changing that to `Stringify`: * `StringAbslStringifySink` → `StringStringifySink` * `OStreamAbslStringifySink` → `OStreamStringifySink` * `WriterAbslStringifySink` → `WriterStringifySink` * `AbslStringifyOStream` → `StringifyOStream` * `AbslStringifyWriter` → `StringifyWriter`
Rationale: `AbslStringify()` has `Absl` prefix so that function names matched
by ADL do not conflict with unrelated functions. In the context of Abseil and
Riegeli, the name "stringify" alone is clear enough.

Remove `UncheckedArrayAbslStringifySink` for now. It will be done differently.

Simplify `riegeli::Join()`. Do not bother with moving instead of copying, and in
consequence do not bother with distinguishing whether the collection is owned or
not. Store `Src` directly instead of `Dependency<const void*, Src>`, which was
used for detecting ownership.

In `WriteWithPadding()`, change the `length` parameter from `size_t` to
`Position`.

Cosmetic in comments.

PiperOrigin-RevId: 781514379